TAS-936 Fixed identification of read-end cut site
The index of the likely read end cut site was not identified when it was
not the first option in the array of likely read ends.  This resulted in
both the cut site remmants and all other data beyond remaining as part
of the tag, thus preventing some tags from aligning properly with the
exernal aligners.